sync + ingest refactor
